requirements.txt
gradio==5.29.0
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
matplotlib==3.10.3
tensorflow==2.12.0
insightface==0.7.3
scikit-learn==1.5.2 